move function that ensures the property `colorLayerOrder`
